pull out auth window code into util class
